output: fix watch mode to respect --units flag
The --watch flag was ignoring the --units option and always displaying
memory values in the default format (MB). This made it inconsistent
with the final output, which correctly honored the --units flag.

Fix this by propagating the units option through to the
RealtimeDisplay struct, allowing it to format memory values according
to the user's preference.

Example of the bug:
  $ peak-mem --units B -w -- sleep 1
  Current RSS: 2.1 MB | Peak RSS: 2.1 MB
  Current VSZ: 5.7 MB | Peak VSZ: 5.7 MB
  Command: sleep 1
  Peak memory usage: 2064384 B (RSS) / 5722112 B (VSZ)

After fix:
  $ peak-mem --units B -w -- sleep 1
  Current RSS: 2109440 B | Peak RSS: 2109440 B
  Current VSZ: 5726208 B | Peak VSZ: 5726208 B
  Command: sleep 1
  Peak memory usage: 2109440 B (RSS) / 5726208 B (VSZ)

Signed-off-by: Charalampos Mitrodimas <charmitro@posteo.net>